Add files via upload
The model takes a prompt as input and generates any kind of images. The image can be anything living or nonliving. If it is a living thing, it has a background. It also has a color and texture and transparency. An overlay or caption can also be added over the image. The image overlay text background and position of the caption are also customized. The base prompt can be customized.